ingredient component partially done
